+++ Changes.md
@@ -1,5 +1,6 @@
 # master
 - Make core react types builtin instead of requiring a shim file.
+- Fix issue with flow back-end when exporting a function component with polymorphic variables.
 
 # 2.27.0
 - In TypeScript, support --noImplicitAny also for indexing objects. (I.e. without --suppressImplicitAnyIndexErrors).